+++ src/dev/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Timer } from '../lib/Timer'
-import { NotificationBehavior } from '../lib/Notification'
+import { NotificationMode } from '../lib/Notification'
 import { NotificationContextProvider } from '../lib/notificationContext'
 import { randomizeNotifications } from './notificationFactory'
 import { NotificationButtons } from './components/NotificationButtons'
@@ -13,7 +13,7 @@ export function App(): JSX.Element {
   const notifications = randomizeNotifications(20)
   return (
     <NotificationContextProvider timer={new Timer({ fps: 10 })} notifications={notifications}>
-      <NotificationContainer behavior={NotificationBehavior.Default} />
+      <NotificationContainer mode={NotificationMode.Expandable} />
       <NotificationManager />
       <NotificationButtons />
     </NotificationContextProvider>